Basics of message syncing done
on sync, loop over each persona and pull out the messages for it. relies
on persistent challenges. if challenge/response fails, it generates a
new one and tries again. works really well, so far.

once we have the ability to do more than one persona, this may be
unwieldy and require batch-vaildation of challenges. for now, screw it,
this works great.