Refactoring/commenting Protected model
Made some adjustments to the PRotected model's code: reformatted some
redundant blocks, tabbed a lot of assignments, and moved some functions
around.

Also made an effort to really go through everything and comment the hell
out of it. This model is the backbone of the app, it's got to be
polished.