Removing antiquated profile.data blob
Now, all data is tracked in the models (as it should be) and is no
longer synced from profile.profile_data (except at the very beginning).
this provides not only optimization, but removes problems where when
switching between top-level controllers, synced data is lost.

also fixing some issues with the bookmarklet.