Background vs app
Making some changes here and there that really flesh out the difference
between an "app" instance (ie, a tab opened to show the turtl app) and a
"background" instance: a process that holds the main data of the app and
occasionally opens in panels.

this is done mainly via variables being set by the extension, but also
setup.js now has a number of mods to make profile loading from the
background into the app smooth.