Implementing new serialization format
issue #52, in the flesh.

tcrypt now returns raw encryptions strings instead of base64 encoding
everything. base64 encoding is the responsibility of the one calling
tcrypt.

the format encodes the version, the description of the payload
(aes/twofish, CBC, padding, etc), the binary IV, and the payload itself.
it allows for variable length payload description, making it
forward-expandable. it is self-describing, so tcrypt will always be able
to decrypt any message in the new format, no matter what cipher or
padding mode etc is used, as long as the key is correct, and with no
extra metadata. just the payload itself.

this format is MUCH more compact, and allows for binary storage.

note that it's NOT implemented anywhere in the app.