display updates to notes
when a new note is added with a file attached, the note *immediately*
renders in the view. note that the note is NOT saved, it only exists in
memory (until the file finishes saving, then the note is really saved).

also, when a file is encrypting, it now shows it as such in the note
preview.

a lot less confusing than before.