Converting everything to use promises
This goes for API calls, model functions, threaded (de)serialization,
etc. everything now uses promises for anything async.

this is largely untested and will probably give us growing pains, but
this is The Right Thing and will be so much better in the future.

note that to do this, i searched

    /(success|error|complete):/
    /options\.(success|error|complete)/

and replaced all code as neccesary. i'm almost positive i'm missing a
return here or there, as that was the most common mistake i found myself
making. need to go back and review.