no longer using file hashes as ids
the reason i did this originally was to be able to sniff out duplicate
files, but i never actually built this functionality out. in other
words, if two notes added the same file, then one of the notes changed
its file, the file for the other note would be wiped out by the
clear_files() call. didn't think this through at all. it makes much more
sense to just give each file a regular cid() and be done with it.

this has the added benfit of speeding up file crypto significantly.
after some benchmarking, it turns out doing SJCL's sha256 on 2.5mb takes
around 30s. kind of sad. i'd hate to see it on larger files.