renaming _modules to 0modules
this truly pains me to admit defeat, but the aapt build script for
android is so incredibly shitty that it hardcodes the ignoring of
folders starting with an underscore when building an APK, and there is
no workaround/patch for the windows version.

thus, the only way to get the stupid files into the apk is to rename the
folder with a 0, which achieves the same effect. this works fine but is
ugly-ish.

i don't like bending my project to the will of a broken build tool but
the only way to test turtl on android is to swallow my pride rename the
damn folder.

...but i do so in protest