Fixing some largish issues with (de)serialization
So in the old turtl, we hijacked .toJSON() to perform the crypto for us.
This worked ok, but we had to be careful about where/when it was called.
It did have the one benefit of automatically serializing relational
objects, which up until this commit, was not happening correctly.

In fact, there was a nice big security hole where we call toJSON when
serializing, grab all public fields, and save them into the DB when
done. HOWEVER, note.file is a public field, but note.file is another
Protected model with *private fields in it*. Since .toJSON() doesn't
serialize for us anymore, this private data was getting added, verbatim,
into the local DB (and presumably the live server).

The fix is to loop over all *public* relational objects before
serializing and serialize them directly, and override the public json
value with the resulting encrypted data.

This is working great so far.

I also updated deserialize to do the same, although it's a bit simpler
because we don't rely on the reutrn value to be encrypted-only.