Tons of app sharing updates.
Board invite interfaces are (almost?) done. Mostly polished. Probably
read for release.

The new Sharing tab is done as well, lets you add invites via code, but
also shows invites synced via the api as well as boards actively being
shared with you. Invites can be (unlocked and) accepted, shares can be
left.

The real hurdles left are mainly parent/child boards and also removing
the correct notes when a board is unshared BUT this logic is actually
better suited to the API, and anything that needs to be added/removed
can be done using the sync system. brilliant.