Add a minify target to Makefile (concatenate js files)
Chrome seems to load and eval js files to quickly and this ends to
evaluating js files before their dependencies, leading to errors like
`FormController is not defined`.

Concatenating js files forces the order of evaluation and solves the
problem. The concatenated turtl.min.js is heavy (about 2Mo) but after
the first load, this file is cached on the browser.