starting core conversion =]
here it is, forks. login/grabbing profile/searching notes all
implemented off of the new rust core.

the current implementation (which is configurable) uses a websocket
adapter, and the turtl_core/sock crate (which wraps core comm in a
polling websocket layer). this allows us to talk to the turtl core
without having to embed the dll in the browser somehow...REALLY great
for testing.

of course, as things move forward, there are some things in the core
that needed changing, but mostly minor. in fact, it's all just kind of
working beautifully so far.

i'm adding a big ass TODO list that should *HOPEFULLY* be somewhat
complete. there's a lot there, but TBH most of it is hooking up wires at
this point...the bulk of the work, in the core, is done. and a lot of
wires got hooked up tonight:

- login works
- loading the profile (spaces/boards/invites)
- searching notes works
- core <--> ui events are acknowledged (with some being wired up)
- wired up model auto-syncing, meaning when a `sync:update` ui event
  comes through, the models/collections update themselves. haven't had a
  chance to test this yet, since writing data isn't baked in yet. soon.

there's some search stuff broken since the conversion started, but it's
in the TODO.

overall, a very successful first run at wiring up the core. the nice
thing about the current architecture is that we can load in/swap in
different core adapters (ie, a java one or a nodejs one) and specify it
in the config. great for going from dev -> production.